+++ second-hand-b/src/main/java/com/codesquad/secondhand/application/service/in/S3ImageService.java
@@ -1,7 +1,7 @@
 package com.codesquad.secondhand.application.service.in;
 
-import com.codesquad.secondhand.adapter.out.s3.S3ImageManager;
-import com.codesquad.secondhand.application.port.ImageUseCase;
+import com.codesquad.secondhand.adapter.out.s3.S3StorageService;
+import com.codesquad.secondhand.application.port.in.ImageUseCase;
 import com.codesquad.secondhand.application.port.in.response.ImageUploadResponse;
 import com.codesquad.secondhand.application.port.out.ImageRepository;
 import com.codesquad.secondhand.domain.image.Image;
@@ -14,13 +14,13 @@ import org.springframework.web.multipart.MultipartFile;
 @Service
 public class S3ImageService implements ImageUseCase {
 
-    private final S3ImageManager s3ImageManager;
+    private final S3StorageService s3StorageService;
     private final ImageRepository imageRepository;
 
     @Transactional
     @Override
     public ImageUploadResponse upload(MultipartFile file) {
-        String imgUrl = s3ImageManager.upload(file);
+        String imgUrl = s3StorageService.upload(file);
         Image savedImage = imageRepository.save(new Image(imgUrl));
         return new ImageUploadResponse(savedImage.getId(), savedImage.getUrl());
     }
@@ -31,6 +31,6 @@ public class S3ImageService implements ImageUseCase {
         Image image = imageRepository.findById(id)
                 .orElseThrow();
         imageRepository.deleteById(id);
-        s3ImageManager.delete(image.getUrl());
+        s3StorageService.delete(image.getUrl());
     }
 }
